Dynamic Data Futures 0.1.02 Changes (2008-07-02) ------------------------------------------------
- Modified ForeignKey_Edit field template to display a [Not Set] entry for required columns
in insert mode. Added a RequiredFieldValidator to ensure that an entry is selected if the
column is required.

- Added Url and EmailAddress field templates.

- Added DynamicHyperLink control. It generates links to Dynamic Data tables for
all actions and works both in data-binding and simple scenarios.

- Modified routes in Global.asax to generate pretty URLs using the new PrettyDynamicDataRoute.

- Changed FilterAttribute.Order property to match the behavior of ColumnOrderAttribute.Order
property. Now each filter has a default Order of 0; negative numbers can be used to push
filters to the fron, while positive numbers can be used to push filters to the back.

git-svn-id: https://aspnet.svn.codeplex.com/svn@11287 e17a0e51-4ae3-4d35-97c3-1a29b211df97